pyproject.toml
[tool.poetry]
name = "aniblog"
version = "0.1.0"
description = ""
authors = ["Dzmitry Zhybryk"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"
orm = "^0.3.1"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
